Don't refactor the HeightMapTools, don't hard code parameters for the cell width
