+++ ihmc-footstep-planning/src/test/java/us/ihmc/footstepPlanning/swing/CollisionFreeSwingCalculatorLogViewer.java
@@ -17,7 +17,7 @@ import us.ihmc.idl.IDLSequence;
 import us.ihmc.robotics.robotSide.SideDependentList;
 import us.ihmc.sensorProcessing.heightMap.HeightMapData;
 import us.ihmc.sensorProcessing.heightMap.HeightMapMessageTools;
-import us.ihmc.perception.gpuHeightMap.HeightMapTools;
+import us.ihmc.sensorProcessing.heightMap.HeightMapTools;
 import us.ihmc.simulationconstructionset.Robot;
 import us.ihmc.simulationconstructionset.SimulationConstructionSet;
