+++ ihmc-footstep-planning/src/visualizers/java/us/ihmc/footstepPlanning/bodyPath/AStarBodyPathSmootherVisualizer.java
@@ -21,7 +21,7 @@ import us.ihmc.log.LogTools;
 import us.ihmc.pathPlanning.HeightMapDataSetName;
 import us.ihmc.sensorProcessing.heightMap.HeightMapData;
 import us.ihmc.sensorProcessing.heightMap.HeightMapMessageTools;
-import us.ihmc.perception.gpuHeightMap.HeightMapTools;
+import us.ihmc.sensorProcessing.heightMap.HeightMapTools;
 import us.ihmc.simulationconstructionset.Robot;
 import us.ihmc.simulationconstructionset.SimulationConstructionSet;
