+++ ihmc-high-level-behaviors/src/main/java/us/ihmc/behaviors/tools/perception/AlternateHeightMapUpdater.java
@@ -10,7 +10,7 @@ import us.ihmc.euclid.referenceFrame.ReferenceFrame;
 import us.ihmc.euclid.tuple3D.Point3D;
 import us.ihmc.sensorProcessing.heightMap.HeightMapManager;
 import us.ihmc.sensorProcessing.heightMap.HeightMapParameters;
-import us.ihmc.perception.gpuHeightMap.HeightMapTools;
+import us.ihmc.sensorProcessing.heightMap.HeightMapTools;
 
 import java.util.Arrays;
 import java.util.concurrent.atomic.AtomicInteger;
