+++ ihmc-high-level-behaviors/src/main/java/us/ihmc/perception/HumanoidActivePerceptionModule.java
@@ -17,7 +17,7 @@ import us.ihmc.log.LogTools;
 import us.ihmc.perception.parameters.PerceptionConfigurationParameters;
 import us.ihmc.perception.tools.PerceptionDebugTools;
 import us.ihmc.ros2.ROS2Node;
-import us.ihmc.perception.gpuHeightMap.HeightMapTools;
+import us.ihmc.sensorProcessing.heightMap.HeightMapTools;
 
 public class HumanoidActivePerceptionModule
 {
