+++ ihmc-sensor-processing/src/main/java/us/ihmc/sensorProcessing/heightMap/HeightMapManager.java
@@ -7,7 +7,6 @@ import us.ihmc.euclid.tuple2D.interfaces.Point2DReadOnly;
 import us.ihmc.euclid.tuple3D.Point3D;
 import us.ihmc.euclid.tuple3D.interfaces.Point3DReadOnly;
 import us.ihmc.log.LogTools;
-import us.ihmc.perception.gpuHeightMap.HeightMapTools;
 
 import java.util.Arrays;
